+++ components/location/card/index.jsx
